fix: support CGO_ENABLED=0 builds with no-op audio stub
Split audio package into build-tag-guarded files:
- audio_cgo.go: real implementation (requires CGO + platform audio libs)
- audio_stub.go: no-op stub for CGO_ENABLED=0 cross-compilation

Also install libasound2-dev in CI test job for Linux ALSA support.
