+++ .github/workflows/ci.yml
@@ -29,6 +29,8 @@ jobs:
       - uses: actions/setup-go@v5
         with:
           go-version-file: go.mod
+      - name: Install ALSA headers
+        run: sudo apt-get update && sudo apt-get install -y libasound2-dev
       - run: go test ./... -v -race
 
   build:
