+++ internal/audio/audio_cgo.go
@@ -1,3 +1,5 @@
+//go:build cgo
+
 package audio
 
 import (
